fix: separate auth logic
